% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_building.R
\name{add_building}
\alias{add_building}
\title{Add building}
\usage{
add_building(data_table = NULL, building = "Bld_1", source_coords = c(0,
  0), dist_from_source = 20, angle_from_source = 0, elev = 0,
  height = 10, length = 5, width = 10, bld_rotation = 0,
  angle_units = "degrees")
}
\arguments{
\item{building}{Building names. Separate multiple sources with commas.}
}
\description{
Create building parameters for BPIP.
}
\examples{
add_building(building = "Bld_1")
}
\keyword{bpip}
\keyword{building}
\keyword{coords}
\keyword{input}

