% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bld_coords.R
\name{bld_coords}
\alias{bld_coords}
\title{Generate building coordinates}
\usage{
bld_coords(source_coords = c(0, 0), dist_from_source = 20,
  angle_from_source = 0, length = 5, width = 10, bld_rotation = 0,
  angle_units = "degrees", show_plot = TRUE)
}
\arguments{
\item{building}{Building names. Separate multiple sources with commas.}
}
\description{
Calculate building coordinates for BPIP.
}
\examples{
bld_coords(building = "Bld_1")
}
\keyword{bpip}
\keyword{building}
\keyword{coords}
\keyword{input}

