% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bld_tbl.R
\name{bld_tbl}
\alias{bld_tbl}
\title{Building options}
\usage{
bld_tbl(prj_title = "BPIP input", building = "Bld_1", height = 10,
  length = 5, width = 10, bld_rotation = 0, angle_units = "degrees",
  elev = 0, n_tiers = 1, bld_xcoords = c(), bld_ycoords = c(),
  dist_from_source = 20, angle_from_source = 0, source_name = "Stack_1",
  source_coords = c(0, 0), source_elev = 0, source_height = 10)
}
\arguments{
\item{building}{Building names. Separate multiple sources with commas.}

\item{dist_from_source}{Distance from center of source to the center of the building.}
}
\description{
Create an input table of BPIP building options.
}
\examples{
bld_tbl(building = "Bld_1")
}
\keyword{bpip}
\keyword{building}
\keyword{coords}
\keyword{input}

