% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/building_tbl.R
\name{building_tbl}
\alias{building_tbl}
\title{Building options}
\usage{
building_tbl(BUILDING = "Bld_1", HEIGHT = "", ELEV = "", XCOORDS = "",
  YCOORDS = "")
}
\arguments{
\item{BUILDING}{Building names. Separate multiple sources with commas.}
}
\description{
Create an input table of BPIP building options.
}
\examples{
building_tbl(BUILDING = "Bld_1")
}
\keyword{bpip}
\keyword{building}
\keyword{coords}
\keyword{input}

