% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downwash_tbl.R
\name{downwash_tbl}
\alias{downwash_tbl}
\title{Downwash options}
\usage{
downwash_tbl(stack = "STK_1", angle = seq(10, 360, 10), buildhgts = 10,
  buildwids = 10, buildlens = 10, xbadj = 1, ybadj = 1)
}
\arguments{
\item{stack}{Stack names.}
}
\description{
Create an input table of downwash options for AERMOD.
}
\examples{
downwash_tbl(stack = "STK_1")
}
\keyword{aermod}
\keyword{building}
\keyword{downwash}
\keyword{input}

