% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw.R
\name{fw}
\alias{fw}
\title{Fixed width strings}
\usage{
fw(string = list("2short", "cutThisString"), nchars = 10)
}
\arguments{
\item{string}{Text to . Can be passed a list.}

\item{nchars}{Number of characters to make each string.}
}
\description{
Adjust the length of a character string by cutting if too long, or by adding spaces to the end if too short.
}
\examples{
fw("cutThisString", width = 10)
}
\keyword{cut}
\keyword{fixed}
\keyword{fw}
\keyword{string}
\keyword{width}

