% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_bpip.R
\name{install_bpip}
\alias{install_bpip}
\title{Install BPIP}
\usage{
install_bpip(dir = "bpip")
}
\arguments{
\item{dir}{Folder for installation}
}
\description{
Download bpip-prime from EPA and install to entered folder.
}
\examples{
install_bpip(dir = "bpip")
}
\keyword{bpip}
\keyword{building}
\keyword{downwash}

