% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_bpip.R
\name{new_bpip}
\alias{new_bpip}
\title{Build a BPIP data frame of building parameters}
\usage{
new_bpip(prj_title = "BPIP input", building = "Bld_1", height = 10,
  width_x = 5, length_y = 10, bld_rotation = 0, angle_units = "degrees",
  elev = 0, n_tiers = 1, bld_xcoords = c(), bld_ycoords = c(),
  dist_from_source = 20, angle_from_source = 0, source_name = "Stack_1",
  source_xy = c(0, 0), source_elev = 0, source_height = 10)
}
\arguments{
\item{building}{Building IDs or names.}

\item{dist_from_source}{Distance from center of source to center of building.}
}
\description{
Create an input table of building parameters for BPIP.
}
\examples{
new_bpip(building = "Boiler house")
}
\keyword{bpip}
\keyword{building}
\keyword{input}

