% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bpip_inp.R
\name{read_bpip_inp}
\alias{read_bpip_inp}
\title{Read BPIP input file}
\usage{
read_bpip_inp(file)
}
\arguments{
\item{file}{A path to a file or a text string containing line breaks. Default is "bpip.inp".}
}
\description{
Read a bpip.inp file into a dataframe.
}
\examples{
bpip_inp <- "'Example1'
'p'
'METERS' 1.00
'UTMN' 0.00
1
'Bld_1' 1 0
4 10
-22.5 5
-17.5 5
-17.5 -5
-22.5 -5
1
'Stack_1' 0 10 0 0
'Stack_2' 0 10 0 0
"

read_bpip_inp(file = bpip_inp)
}
\keyword{bpip}
\keyword{input}
\keyword{read}
