% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bpip_out.R
\name{read_bpip_out}
\alias{read_bpip_out}
\title{Read BPIP output file}
\usage{
read_bpip_out(file = "bpip.out", as_text = TRUE)
}
\arguments{
\item{file}{File location. Default is "bpip.out" in the working directory.}

\item{as_text}{Output results as a character string. If FALSE ouput as a dataframe.}
}
\description{
Read a bpip.out file into a dataframe.
}
\examples{
read_bpip_out(file = "bpip.out")
}
\keyword{bpip}
\keyword{output}
\keyword{read}
\keyword{results}

