% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_bld.R
\name{rotate_bld}
\alias{rotate_bld}
\title{Rotated rectangle coordinates}
\usage{
rotate_bld(center_x = 0, center_y = 0, width_x = 4, length_y = 3,
  rotation = 0, angle_units = "degrees", show_plot = TRUE)
}
\arguments{
\item{center_x}{Building's center X coordinate.}

\item{center_y}{Building's center Y coordinate.}
}
\description{
Generate coordinates for a rotated rectangle.
}
\examples{
rotate_bld(center_x = 0, center_y = 0, width_x = 2, length_y = 1)
}
\keyword{bpip}
\keyword{building}
\keyword{coords}
\keyword{rectangle}

