% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bpip.R
\name{run_bpip}
\alias{run_bpip}
\title{Run BPIP}
\usage{
run_bpip(data = bpip_inp, output = "bpip_results", exe_folder = getwd())
}
\arguments{
\item{data}{Dataframe of bpip parameters.}

\item{output}{Filename for bpip results.}

\item{exe_folder}{Folder containing bpip.exe.}
}
\description{
Call BPIP.exe and run input file
}
\examples{
run_bpip(data = bpip_inp, output = "bpip_results")
}
\keyword{bpip}
\keyword{building}
\keyword{coords}
\keyword{rectangle}

