% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_building.R
\name{add_building}
\alias{add_building}
\title{Add building}
\usage{
add_building(BUILDING = "Bld_1", SOURCE_COORDS = c(0, 0), DISTANCE = 20,
  ANGLE_DEGREES = 10, ELEV = 1, HEIGHT = 10, LENGTH = 5, WIDTH = 10)
}
\arguments{
\item{BUILDING}{Building names. Separate multiple sources with commas.}
}
\description{
Create building parameters for BPIP.
}
\examples{
add_building(BUILDING = "Bld_1")
}
\keyword{bpip}
\keyword{building}
\keyword{coords}
\keyword{input}

