% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bpip.R
\name{plot_bpip}
\alias{plot_bpip}
\title{Plot buildings from a BPIP dataframe}
\usage{
plot_bpip(data)
}
\arguments{
\item{data}{Data frame of bpip parameters. Run 'bld_tbl()' for an example.}
}
\description{
Plot a BPIP building dataframe.
}
\examples{
bpip_blds <- bld_tbl()

plot_bpip(data = bpip_blds)
}
\keyword{bpip}
\keyword{buildings}
\keyword{downwash}
\keyword{plot}

