% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bpip_inp.R
\name{read_bpip_inp}
\alias{read_bpip_inp}
\title{Read BPIP input file}
\usage{
read_bpip_inp(file = "bpip.inp")
}
\arguments{
\item{path}{File location. Default is "bpip.inp" in the working directory.}
}
\description{
Read a bpip.inp file into a dataframe.
}
\examples{
read_bpip_inp(file = "bpip.inp")
}
\keyword{bpip}
\keyword{input}
\keyword{read}

