% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangle.R
\name{rectangle}
\alias{rectangle}
\title{Rectangle coordinates}
\usage{
rectangle(center_x = 0, center_y = 0, length = 3, width = 4,
  rotation = 0, angle_units = "degrees", show_plot = TRUE)
}
\arguments{
\item{center_x}{Building's center X coordinate.}

\item{center_y}{Building's center Y coordinate.}
}
\description{
Generate rectangle coordinates for BPIP buildings.
}
\examples{
rectangle(center_x = 0, center_y = 0, length = 1, width = 2)
}
\keyword{bpip}
\keyword{building}
\keyword{coords}
\keyword{rectangle}

