% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bpip.R
\name{run_bpip}
\alias{run_bpip}
\title{Run BPIP}
\usage{
run_bpip(input = "bpip.inp", output = "bpip_results",
  exe_folder = "bpip_exe")
}
\arguments{
\item{input}{Path to input file.}

\item{output}{Filename for bpip results.}

\item{exe_folder}{Folder containing bpip.exe.}
}
\description{
Call BPIP.exe and run input file
}
\examples{
run_bpip(data = "bpip.inp", 
         output = "bpip_results", 
         exe_folder = "bpip_exe")
}
\keyword{bpip}
\keyword{building}
\keyword{coords}
\keyword{rectangle}

