% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bpip.R
\name{run_bpip}
\alias{run_bpip}
\title{Run the BPIP downwash model}
\usage{
run_bpip(input = "bpip.inp", output = "bpip_results", exe_folder = "bpip")
}
\arguments{
\item{input}{Path to input file.}

\item{output}{Filename for BPIP results.}

\item{exe_folder}{Folder containing bpip.exe.}
}
\description{
Call bpip.exe and run input file.
}
\examples{
\dontrun{
run_bpip(input      = "bpip.inp", 
         output     = "bpip_results", 
         exe_folder = "EPA/bpip")
}
}
\keyword{bpip}
\keyword{building}
\keyword{downwash}
