% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bpip.R
\name{write_bpip}
\alias{write_bpip}
\title{Write building table to a BPIP input file.}
\usage{
write_bpip(data, path = "bpip.inp", prj_title = gsub("[.]inp", "", path),
  output_type = "p")
}
\arguments{
\item{data}{Dataframe of bpip parameters.}

\item{path}{Path to write to. Default is "bpip.inp".}

\item{prj_title}{Title of project added to BPIP file. Defaults to path name.}

\item{output_type}{A character flag controls the BPIP model's output type.
The three flags are:
    (1) 'p' for PRIME or AERMOD input;  
    (2) 'ST' for ISCST2 input; 
    (3) 'LT' for ISCLT2 input.}
}
\description{
Output a BPIP input file from a data frame of building parameters.
}
\examples{
bpip_blds <- bld_tbl()

write_bpip(data = bpip_blds)
}
\keyword{bpip}
\keyword{building}
\keyword{input}
\keyword{save}
\keyword{write}

