% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bpip.R
\name{write_bpip}
\alias{write_bpip}
\title{Write building table to a BPIP input text file.}
\usage{
write_bpip(data, path = "bpip.inp", prj_title = gsub("[.]inp", "", path),
  output_type = "p")
}
\arguments{
\item{data}{Dataframe of bpip parameters.}

\item{path}{Path to write to. Default is "bpip.inp".}

\item{prj_title}{Title of project added to BPIP file. Defaults to path name.}

\item{output_type}{A character flag controls the BPIP model's output type. For input to either an ISCST2 or an ISCLT2 input file. 
The three flags are:
    'p'  - Output for PRIME or AERMOD input. 
    'ST' - Output for ISCST2 input.
    'LT' - Output for ISCLT2 input.}
}
\description{
Output a BPIP input file from a data frame of building parameters.
}
\examples{
write_bpip(data = facility_blds)
}
\keyword{bpip}
\keyword{building}
\keyword{input}
\keyword{save}
\keyword{write}

